+++ packages/cli/src/pywrangler/utils.py
@@ -95,7 +95,7 @@ def setup_logging() -> int:
     return log_level
 
 
-def _get_pywrangler_version() -> str:
+def get_pywrangler_version() -> str:
     """Get the version of pywrangler."""
     try:
         from importlib.metadata import version
@@ -109,7 +109,7 @@ def log_startup_info() -> None:
     """
     Log startup information for debugging.
     """
-    logger.debug(f"pywrangler version: {_get_pywrangler_version()}")
+    logger.debug(f"pywrangler version: {get_pywrangler_version()}")
     logger.debug(f"Python: {platform.python_version()}")
     logger.debug(f"Platform: {sys.platform}")
     logger.debug(f"Working directory: {Path.cwd()}")
@@ -390,6 +390,14 @@ def get_uv_pyodide_interp_name() -> str:
 
 
 def get_pyodide_index() -> str:
+    """
+    Get the Pyodide index URL based on the Python version.
+    This doesn't have to match the Pyodide version that Python workers use,
+    as long as they are ABI compatible.
+
+    When updating this, run scripts/compare_pyodide_index.py to see if there are missing
+    packages in the index.
+    """
     match get_python_version():
         case "3.12":
             v = "0.27.7"