+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,54 @@
 
 <!-- version list -->
 
+## v1.11.0 (2026-06-03)
+
+### Features
+
+- Reverts #117
+  ([`973379e`](https://github.com/cloudflare/workers-py/commit/973379ea67188275f28387b7d22d5bbdcf61fe06))
+
+
+## v1.10.0 (2026-06-02)
+
+### Features
+
+- Update Python 3.13 package index from Pyodide 0.28.3 to 0.29.3
+  ([#117](https://github.com/cloudflare/workers-py/pull/117),
+  [`9c322af`](https://github.com/cloudflare/workers-py/commit/9c322af22c29223d2452fa532cfeae27d9ee0767))
+
+
+## v1.9.5 (2026-06-01)
+
+### Bug Fixes
+
+- Don't install Pyodide interpreter onto user's PATH
+  ([#115](https://github.com/cloudflare/workers-py/pull/115),
+  [`e6e78c4`](https://github.com/cloudflare/workers-py/commit/e6e78c4f22dd1219ac1e4c83cb57f3cb2b997d9b))
+
+
+## v1.9.4 (2026-05-19)
+
+### Bug Fixes
+
+- Wrap DurableObject.abort() so that python cleanup can be done before abort
+  ([#106](https://github.com/cloudflare/workers-py/pull/106),
+  [`bf6acf2`](https://github.com/cloudflare/workers-py/commit/bf6acf24429fb1525f34334ff2cefffa45b287ef))
+
+- **runtime-sdk**: Wrap DO.abort() to cleanup stale tasks before abortion
+  ([#106](https://github.com/cloudflare/workers-py/pull/106),
+  [`bf6acf2`](https://github.com/cloudflare/workers-py/commit/bf6acf24429fb1525f34334ff2cefffa45b287ef))
+
+
+## v1.9.3 (2026-04-22)
+
+### Bug Fixes
+
+- **workers-py**: Bust cache if workers-py version changes
+  ([#95](https://github.com/cloudflare/workers-py/pull/95),
+  [`9804681`](https://github.com/cloudflare/workers-py/commit/9804681956585c6e017aeda53924561116dccaf7))
+
+
 ## v1.9.2 (2026-04-02)
 
 ### Bug Fixes